moving loss-rate tests to a separate file
